tests/wpt/web-platform-tests/largest-contentful-paint/observe-image.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Largest Contentful Paint: observe image.</title>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  async_test(function (t) {
    if (!window.LargestContentfulPaint) {
      assert_unreached("LargestContentfulPaint is not implemented");
    }
    let beforeRender = performance.now();
    const observer = new PerformanceObserver(
      t.step_func_done(function(entryList) {
        assert_equals(entryList.getEntries().length, 1);
        const entry = entryList.getEntries()[0];
        assert_equals(entry.entryType, 'largestContentfulPaint');
        assert_greater_than_equal(entry.startTime, beforeRender,
          'The rendering timestamp should occur after script starts running.');
        assert_greater_than_equal(performance.now(), entry.startTime,
          'The rendering timestamp should occur before the entry is dispatched to the observer.');
        assert_equals(entry.duration, 0);
        // blue.png is 133 x 106.
        assert_equals(entry.size, 14098);
      })
    );
    observer.observe({type: 'largestContentfulPaint', buffered: true});
  }, 'Element with elementtiming attribute is observable.');
</script>

<img src='/images/blue.png'/>
</body>
